Software T&C/Acknowledgement: Add link to T&C and disclaimer
We want to pop up a modal box to tell users they need to be careful when
inscribing.  It's not meant to scare them, but realistically the <frame>
tag we are previewing in is not going to be an exact replica of every
potential viewer out there (e.g., ord.io).

Note that the acknowledgement recurs every 2 weeks, though that is
likely more frequently than desired.

[ Testing: npm run dev ]