OrdinalsBot: Submit to the /orders API (no response displayed to user)
This change merely submits to the /orders API and retrieves a response,
but does no rendering (future exercise).

Full docs: https://docs.ordinalsbot.com/#create-order

NOTE: Because we use base64 encoding we have to add a prepended data:
scheme along with the MIME-type because the file name is essentially
ignored by the API.

NOTE: The referral code is invalid as of this commit.

[ Testing: npm run dev, inspected console logs ]