XVerse/Unisat: Add wallet connectors for Chrome extensions
The detailed docs can be found here:
https://docs.xverse.app/sats-connect
https://docs.unisat.io/dev/unisat-wallet-api

Note that we cannot do the effects in the rendering by calling it (as if
it were a simple .enable()) because XVerse appears to be reconnecting
and re-asking the user every time we request the wallet address.  This
may just be a localhost thing.

[ Testing: npm run dev ]