iframe: Use data:text/html... URLs for iframe
Documentation shows that the <frame> tag is not supported in HTML5 (and
we are already seeing issues with rendering and updating anyway.  This
change simply moves toward an iframe with an `src` that is constantly
changing.  By base64-encoding and specifying a mime type, we can use the
<iframe> component and constantly re-render.

NOTE: We have to move from ord.io to magiceden due to CORS policies

[ Testing: npm run build, npm run dev ]