SVG Support: Split out SVG vs. HTML encodings
Before, we were just using one CodeMirror object and changing the
extensions type.  Now, we use a generic `codeValue` which can be
encapsulated with different values in `orderData` depending on the
`mimeType` attribute.

As we expand support for other protocols (e.g., layered images), we will
continue to expand the number of supported items in teh orderData.

[ Testing: npm run build, npm run dev ]