Sweep Prototype: MagicEden Sweep attempt (failed)
This commit includes a page at /sweep that mocks out what it would take
to try to sweep multiple inscriptions from MagicEden.

However, we discovered in doing this that it's not possible because the
endpoint at `/ord/btc/psbt/buying` returns a fully unsigned PSBT and
sends back a POST request to the MagicEden server at the very end for
combining the PSBTs.

This negates the decentralization of signing it in the client side and
broadcasting it to the network.

[ Testing: npm run build, npm run dev ]