feat(Image Editing): Support editing images that are already inscribed
The basic idea here is to load a source from the inscription ID that a
user specifies and give it to the filerobot-image-editor widget
(https://github.com/scaleflex/filerobot-image-editor).

The default image and inscription is, of course, the Clay Tangz.

NOTE: There are some weird workarounds we have to do with dynamic
imports since FIE relies so heavily on the `window` and `canvas`
objects.

[ Testing: npm run build, npm run dev, manual QA ]