feat(Collection Discounts): Enable discounts for the dev fee for certain collections
Previously, we were charging a flat 4k sats per inscription.  However,
we occassionally will want to discount this utility (e.g., for the
Otternals).  So, we add an environment variable to detect a list of URLs
where collections are housed (conforming to the following ME spec):
https://docs.magiceden.io/reference/gettokens-1

NOTE: We add a little disclaimer on the footer to make the pricing
slightly clearer.

[ Testing: npm run build, manual QA ]